fix: filters oracle notifications per space
Previous implentation worked under the asumption that the oracle contract was
unique per space, but it is not. To identify if a answer event is related to
a subscribed space, the application needs to keep track of the existing
proposals and their `questionId`.

This commit includes the following changes:
- Code for processing proposals and answers is extracted into their own
functions.
- Once a proposal for a subscribed space is detected, it is recorded in the db.
- As the proposal (more specifically, the `questionId`) is required to correctly
identify an answer, proposals and answers are no longer processed concurrently
but sequentially.
- A `happenedAt` field has been included in both proposals and answers. This
will allow ignoring stale events and recorded proposals.
- Adjusted tests to make use of the random generators.